ai-ethics/requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.20
langgraph==0.0.20
openai==1.10.0
python-dotenv==1.0.0
faiss-cpu==1.7.4
pypdf==4.0.0
tavily-python==0.3.0
pydantic==2.5.0